scraper/requirements.txt
# Python 3.8+ required
# Recommended: Python 3.9 or 3.10

# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
supabase==2.0.2
python-dotenv==1.0.0
pydantic==2.5.0
lxml==4.9.3
fake-useragent==1.4.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.3.1
